requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
mongoengine==0.15.3
msgpack==0.5.6
pycryptodome==3.6.4
pyeclib==1.5.0
pymongo==3.7.1
pytz==2018.5
pyzmq==17.1.0
redis==2.10.6
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.10
urllib3==1.23
uuid==1.30
zmq==0.0.0
